transcode-video-to-audio/requirements.txt
# Copyright 2016 Google Inc. All Rights Reserved.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.


Flask==0.10.1
requests==2.18.4
google-cloud-storage==1.6.0
urllib3==1.22
gunicorn==19.6.0
PyMySQL==0.7.3
six==1.10.0
honcho==0.7.1
oauth2client==4.1.2